services/backend/requirements/install.txt
# =============================================================
# requirements/install.txt -- backend service (pip format)
# =============================================================

fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.0.1
pandas
openpyxl
pyexcel
pyexcel-xls
PyPDF2
python-multipart
httpx==0.27.0
bcrypt
PyJWT
pika
